Track all TimelineElement in a tree
This way we can more easily do operation with them.

This commit also introduces restrictions when two elements
fully overlap, this is not possible anymore.

Modifies all test to take that new restriction into account.

The plan is to move all the TimelineEdition to using that data
structure, which should simplify immensly the code.